requirements.txt
# === Core ===
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5
python-dotenv==1.1.1
requests==2.32.5
tqdm==4.67.1

# === Web Framework ===
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.9

# === LangChain + OpenAI (versioni aggiornate per compatibilità con langgraph 1.0) ===
langchain>=0.3.9
langchain-core>=1.0.0
langchain-openai>=0.3.3
langchain-community>=0.3.7
langchain-text-splitters>=0.3.2
langsmith>=0.1.145

# === LangGraph ===
langgraph==1.0.2
langgraph-checkpoint>=2.0.4
langgraph-prebuilt>=1.0.2

# === OpenAI ===
openai>=1.58.1,<2.0.0
tiktoken>=0.8.0,<1.0.0

# === Vector DB / Embeddings ===
faiss-cpu==1.12.0

# === Utility ===
regex==2024.11.6
PyYAML==6.0.2

# === Cryptography ===
pycryptodome==3.20.0

# === Word Document Parser ===
python-docx==1.1.2

# === Models / Validation ===
pydantic==2.9.2